Move connect logic to peer manager
